Reworked quantity selector to be more dry by creating a quant dropdown component
